requirements.txt
djangorestframework
django-cors-headers

djoser
openpyxl
